test/fixtures/wpt/encoding/bom-handling.html

<!-- starts with a UTF-8 BOM -->
<!doctype html>
<meta charset="windows-1252">
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<title>BOM handling</title>
<div id=log></div>
<script>

test(
  () => {
    assert_equals(document.characterSet.toLowerCase(), "utf-8");
  },
  "document.characterSet should match the BOM"
);

</script>
